+++ pyproject.toml
@@ -42,3 +42,4 @@ dependencies = []
 
 [tool.maturin]
 features = ["pyo3/extension-module"]
+include = ["hl7conv2"]
